Refactor: 거래게시판 리팩토링

+++ src/back-end/timepay/src/main/java/com/capstone/timepay/controller/board/DealBoardController.java
@@ -1,5 +1,6 @@
 package com.capstone.timepay.controller.board;
 
+import com.capstone.timepay.domain.dealBoard.DealBoard;
 import com.capstone.timepay.service.board.dto.DealBoardDTO;
 import com.capstone.timepay.service.board.service.DealBoardService;
 import io.swagger.annotations.ApiOperation;
@@ -7,8 +8,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.data.domain.Page;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RequiredArgsConstructor
 @RestController
@@ -19,9 +23,16 @@ public class DealBoardController
 
     @ApiOperation(value = "거래게시판 모든 게시판 불러오기")
     @GetMapping("")
-    public ResponseEntity<List<DealBoardDTO>> getBoards()
+    public ResponseEntity<Page<DealBoardDTO>> getBoards(
+            @RequestParam(value = "pagingIndex", defaultValue = "0") int pagingIndex,
+            @RequestParam(value = "pagingSize", defaultValue = "50") int pagingSize)
     {
-        return new ResponseEntity(dealBoardService.getDealBoards(), HttpStatus.OK);
+        Page<DealBoardDTO> paging = dealBoardService.getDealBoards(pagingIndex, pagingSize);
+        if (paging.isEmpty())
+        {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(paging, HttpStatus.OK);
     }
 
     @ApiOperation(value = "거래게시판 개별 게시판 불러오기")
@@ -31,43 +42,61 @@ public class DealBoardController
         return new ResponseEntity(dealBoardService.getDealBoard(id), HttpStatus.OK);
     }
 
-    // 숨김처리 안된 게시물 조회
-//    @ResponseEntityStatus(HttpStatus.OK)
-//    @GetMapping("/good")
-//    public ResponseEntity<?> getGoodBoards()
-//    {
-//        return new ResponseEntity("SUCCESS", "숨김처리 안된 게시판 조회", dealBoardService.getGoodBoard());
-//    }
-//
-//
-//    // 숨김처리된 게시물 조회
-//    @ResponseEntityStatus(HttpStatus.OK)
-//    @GetMapping("/bad")
-//    public ResponseEntity<?> getBadBoards()
-//    {
-//        return new ResponseEntity("SUCCESS", "숨김처리 게시판 조회", dealBoardService.getBadBoard());
-//    }
-//
-
     @ApiOperation(value = "거래게시판 게시글 작성")
     @PostMapping("/write")
-    public ResponseEntity<?> write(@RequestBody DealBoardDTO dealBoardDTO)
+    public ResponseEntity write(@RequestBody DealBoardDTO dealBoardDTO)
     {
         return new ResponseEntity(dealBoardService.write(dealBoardDTO), HttpStatus.OK);
     }
 
     @ApiOperation(value = "거래게시판 게시글 수정")
     @PutMapping("/update/{id}")
-    public ResponseEntity<?> update(@RequestBody DealBoardDTO dealBoardDTO, @PathVariable("id") Long id)
+    public Map<String, Object> update(@RequestBody DealBoardDTO dealBoardDTO, @PathVariable("id") Long id)
     {
-        return new ResponseEntity(dealBoardService.update(id, dealBoardDTO), HttpStatus.OK);
+        Map<String, Object> updateMap = new HashMap<>();
+        DealBoard dealBoard = dealBoardService.getId(id);
+        if (dealBoard == null)
+        {
+            updateMap.put("success", false);
+            updateMap.put("message", "해당 게시글을 찾을 수 없습니다");
+            return updateMap;
+        }
+
+        if (!dealBoard.getUuid().equals(dealBoardDTO.getUuid()))
+        {
+            updateMap.put("success", false);
+            updateMap.put("message", "수정 권한이 없습니다");
+            return updateMap;
+        }
+
+        dealBoardService.update(id, dealBoardDTO);
+        updateMap.put("success", true);
+        updateMap.put("dealBoard", dealBoard);
+        return updateMap;
     }
 
     @ApiOperation(value = "거래게시판 게시글 삭제")
     @DeleteMapping ("/delete/{id}")
-    public ResponseEntity<?> delete(@PathVariable("id") Long id)
+    public Map<String, Object> delete(@RequestBody DealBoardDTO dealBoardDTO, @PathVariable("id") Long id)
     {
+        Map<String, Object> deleteMap = new HashMap<>();
+        DealBoard dealBoard = dealBoardService.getId(id);
+        if (dealBoard == null)
+        {
+            deleteMap.put("success", false);
+            deleteMap.put("message", "해당 게시글을 찾을 수 없습니다");
+            return deleteMap;
+        }
+
+        if (!dealBoard.getUuid().equals(dealBoardDTO.getUuid()))
+        {
+            deleteMap.put("success", false);
+            deleteMap.put("message", "삭제 권한이 없습니다");
+            return deleteMap;
+        }
+
         dealBoardService.delete(id);
-        return new ResponseEntity(HttpStatus.OK);
+        deleteMap.put("success", true);
+        return deleteMap;
     }
 }
